requirements.txt
# requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
bardapi==0.1.38
werkzeug==2.3.7
SQLAlchemy==2.0.25
pytest==7.4.3
pytest-flask==1.3.0